feat: compile for go1.20 and go1.21
