+++ logger.go
@@ -2,7 +2,6 @@ package log
 
 import (
 	"bytes"
-	"context"
 	"fmt"
 	"io"
 	"os"
@@ -14,7 +13,6 @@ import (
 
 	"github.com/charmbracelet/lipgloss"
 	"github.com/muesli/termenv"
-	"golang.org/x/exp/slog"
 )
 
 var (
@@ -49,53 +47,6 @@ type Logger struct {
 
 	helpers *sync.Map
 }
-
-// Enabled reports whether the logger is enabled for the given level.
-//
-// Implements slog.Handler.
-func (l *Logger) Enabled(_ context.Context, level slog.Level) bool {
-	return atomic.LoadInt32(&l.level) <= int32(fromSlogLevel[level])
-}
-
-// Handle handles the Record. It will only be called if Enabled returns true.
-//
-// Implements slog.Handler.
-func (l *Logger) Handle(_ context.Context, record slog.Record) error {
-	fields := make([]interface{}, 0, record.NumAttrs()*2)
-	record.Attrs(func(a slog.Attr) bool {
-		fields = append(fields, a.Key, a.Value.String())
-		return true
-	})
-	// Get the caller frame using the record's PC.
-	frames := runtime.CallersFrames([]uintptr{record.PC})
-	frame, _ := frames.Next()
-	l.handle(fromSlogLevel[record.Level], record.Time, []runtime.Frame{frame}, record.Message, fields...)
-	return nil
-}
-
-// WithAttrs returns a new Handler with the given attributes added.
-//
-// Implements slog.Handler.
-func (l *Logger) WithAttrs(attrs []slog.Attr) slog.Handler {
-	fields := make([]interface{}, 0, len(attrs)*2)
-	for _, attr := range attrs {
-		fields = append(fields, attr.Key, attr.Value)
-	}
-	return l.With(fields...)
-}
-
-// WithGroup returns a new Handler with the given group name prepended to the
-// current group name or prefix.
-//
-// Implements slog.Handler.
-func (l *Logger) WithGroup(name string) slog.Handler {
-	if l.prefix != "" {
-		name = l.prefix + "." + name
-	}
-	return l.WithPrefix(name)
-}
-
-var _ slog.Handler = (*Logger)(nil)
 
 func (l *Logger) log(level Level, msg interface{}, keyvals ...interface{}) {
 	if atomic.LoadUint32(&l.isDiscard) != 0 {
